+++ src/navigation/Root.tsx
@@ -3,12 +3,19 @@ import { ColorSchemeName } from 'react-native';
 import { NavigationContainer } from '@react-navigation/native';
 import { navigationLightTheme, navigationDarkTheme } from 'src/constants/theme';
 import { useAppSelector } from 'src/hooks/useful-ducks';
-import { MainStackNavigator } from './main';
 import { AuthStackNavigator } from './auth-stack';
+import { BottomTabNavigator } from './bottom-tab';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
 
 export interface RootParams {
     scheme?: ColorSchemeName; // the color scheme of the app
 }
+
+export type RootStackParams = {
+    Main: undefined
+};
+
+const StackNav = createNativeStackNavigator<RootStackParams>();
 
 export const RootNavigator: React.FC<RootParams> = ({ scheme }) => {
     // redux handlers
@@ -16,7 +23,13 @@ export const RootNavigator: React.FC<RootParams> = ({ scheme }) => {
 
     return (
         <NavigationContainer theme={scheme === 'dark' ? navigationDarkTheme : navigationLightTheme}>
-            {loggedIn ? <MainStackNavigator /> : <AuthStackNavigator />}
+            <StackNav.Navigator>
+                <StackNav.Screen
+                    name='Main'
+                    component={loggedIn ? BottomTabNavigator : AuthStackNavigator}
+                    options={{ headerShown: false }}
+                />
+            </StackNav.Navigator>
         </NavigationContainer>
     );
 };
