+++ src/navigation/home-stack.tsx
@@ -4,10 +4,12 @@ import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import { useNavigation } from '@react-navigation/native';
 import { HomeScreen, PickupScreen } from 'src/screens';
 import { MaterialCommunityIcons } from '@expo/vector-icons';
+import { AuthStackNavigator } from './auth-stack';
 
 export type HomeStackParams = {
     Home: undefined;
     PickupSession: undefined;
+    Auth: undefined;
 };
 
 const StackNav = createNativeStackNavigator<HomeStackParams>();
@@ -31,6 +33,22 @@ export const HomeStackNavigator: React.FC<any> = () => {
         );
     };
 
+    const checkLogin = () => {
+        Alert.alert(
+            'Are you sure you want to exit?',
+            'Your progress will not be saved.',
+            [
+                { text: 'Exit', onPress: navigation.goBack, style: 'destructive' },
+                {
+                    text: 'Return',
+                    onPress: () => null,
+                    style: 'cancel',
+                },
+            ],
+            { cancelable: false },
+        );
+    };
+
     return (
         <StackNav.Navigator>
             <StackNav.Screen name="Home" component={HomeScreen} options={{ headerTitle: 'Home' }} />
@@ -42,6 +60,15 @@ export const HomeStackNavigator: React.FC<any> = () => {
                     headerRight: () => CloseIcon(checkPickup),
                     presentation: 'fullScreenModal',
                     gestureEnabled: false,
+                }}
+            />
+            <StackNav.Screen
+                name="Auth"
+                component={AuthStackNavigator}
+                options={{
+                    headerTitle: 'Login or Sign Up',
+                    headerRight: () => CloseIcon(checkLogin),
+                    presentation: 'modal',
                 }}
             />
         </StackNav.Navigator>
