--- src/navigation/main.tsx
@@ -1,68 +1,0 @@
-import React from 'react';
-import { Alert } from 'react-native';
-import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import { useNavigation } from '@react-navigation/native';
-import { Icon } from 'native-base';
-import { MaterialCommunityIcons } from '@expo/vector-icons';
-import { AuthStackNavigator } from './auth-stack';
-import { BottomTabNavigator } from './bottom-tab';
-
-export type MainStackParams = {
-    Main: undefined;
-    Auth: undefined;
-};
-
-const StackNav = createNativeStackNavigator<MainStackParams>();
-
-const CloseIcon = (onClose: () => void) => (
-    <Icon
-        as={MaterialCommunityIcons}
-        color="plainText.800"
-        name="close"
-        size={22}
-        onPress={onClose}
-    />
-);
-
-export const MainStackNavigator: React.FC<any> = () => {
-    /*
-    This stack navigator will add the login modal that can pop up from wherever in the app, 
-    if the user does certain actions that require an account
-  */
-    const navigation = useNavigation();
-
-    const checkLogin = () => {
-        Alert.alert(
-            'Are you sure you want to exit?',
-            'Your progress will not be saved.',
-            [
-                { text: 'Exit', onPress: navigation.goBack, style: 'destructive' },
-                {
-                    text: 'Return',
-                    onPress: () => null,
-                    style: 'cancel',
-                },
-            ],
-            { cancelable: false },
-        );
-    };
-
-    return (
-        <StackNav.Navigator id="MainStackNavigator">
-            <StackNav.Screen
-                name="Main"
-                component={BottomTabNavigator}
-                options={{ headerShown: false }}
-            />
-            <StackNav.Screen
-                name="Auth"
-                component={AuthStackNavigator}
-                options={{
-                    headerTitle: 'Login or Sign Up',
-                    headerRight: () => CloseIcon(checkLogin),
-                    presentation: 'modal',
-                }}
-            />
-        </StackNav.Navigator>
-    );
-};
